requirements.txt
aiomysql==0.0.22
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.3.0
cffi==1.15.0
click==8.1.0
colorama==0.4.4
cryptography==36.0.2
databases==0.5.5
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==0.9.3
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
SQLAlchemy-serializer==1.4.1
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.6
watchgod==0.8.1
websockets==10.2
zipp==3.8.0
